Extract generators
